修正了背开启值负数时加热不正常的bug 	modified:   config.h 	modified:   gui.c 	modified:   main.c 	modified:   relay.c 	modified:   relay.h

+++ software_src/relay.c
@@ -15,7 +15,7 @@ void relay_init(void){
 *传出 ： 当前继电器状态
 *
 */
-void relay(float temp)
+void relay(signed int temp)
 {
 	if(temp<config.heatThreshold)
 	{
